Elimated a large accumulation of technical debt accross the application

+++ src/main/java/com/dto/CartDTO.java
@@ -1,7 +1,5 @@
 package com.dto;
 
-
-
 import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,8 +13,8 @@ import lombok.NoArgsConstructor;
 public class CartDTO {
     private long cartId;
     private Set<CartItemDTO> cartItems;
-    private int totalItems;
-    private double totalPrice;
-    private int totalDiscountedPrice;
-    private long userId; 
+    private Integer totalItems;
+    private Integer totalPrice;
+    private Integer totalDiscountedPrice;
+    private long userId;
 }
